Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,17 +1,18 @@
 #include<stdio.h>
 int main(){
-    int a,b;
+    float a,b;
     char c;
-    scanf("%d %d ",&a ,&b);
+
+    scanf("%f %f ",&a ,&b);
     scanf("%c",&c);
     if(c=='+'){
-        printf("%d", a+b);
+        printf("%f", a+b);
     }
     else if (c=='-'){
-        printf("%d",a-b);
+        printf("%f",a-b);
     }
     else if (c=='*'){
-        printf("%d",a*b);
+        printf("%f",a*b);
     }
     else if (c=='/'){
         printf("%f",a/b);
